firmware/hab-server/requirements.txt
adafruit-circuitpython-bmp3xx
python-periphery==2.3.0
pyserial
kiss
gpsdclient
pyYAML
pyalsaaudio
git+https://github.com/gitlynx/kiss.git#egg=kiss
git+https://github.com/jj1bdx/aprs-python#egg=aprs
